Fix handling of indices in C interface for SSIDS
C vs Fortran indexing was flipped in the C interface for SSIDS: If the user set `spral_ssids_options::array_base` to `0` to use C indexing, SSIDS would assume the input is using Fortran indexing. If the user set `spral_ssids_options::array_base` to `1` to use Fortran indexing, SSIDS would assume the input is using C indexing. Looks like this was a copy paste error from the C interface of MA97, where indexing is checked with `.not. f_arrays`. Instead of `f_arrays` we have `cindexed` here, so `.not. f_arrays` should have become `cindexed`. But as the `.not.` was not removed the logic became flawed.