Fix n = 0 case for C interface of SSIDS (#83)
A valid C implementation may return NULL for malloc(0). In that sense,
zero-length arrays may be represented by a null pointer in C. When those
are passed to the C interface of SSIDS, they are detected as not
C_ASSOCIATED, and the corresponding Fortran pointers are nullified. A
few of those nullified Fortran pointers were then used illegally without
further checks. To fix this, we avoid checking for C_ASSOCIATED
altogether by declaring the C arguments as arrays directly.