Fix num_factor and num_flops info after factorize
According to the documentation, num_factor and num_flops should hold the
values "without pivoting after analyse phase, with pivoting after
factorize phase". The latter did not seem to work correctly:

* num_factor and num_flops were initialized with the values from akeep,
i.e. with the values from the analyse phase. The actually counted values
during factorization were therefore added on top. That caused the result
to be up to twice as large (exactly twice in the posdef case) of what it
should be. As a fix, num_factor and num_flops are now initialized to 0.
* The CPU factorization did not seem to count these statistics at all.
Together with the issue mentioned before, that caused num_factor and
num_flops to be reported correctly after the factorize phase in the
posdef case. In the indef case, any extra nnz or flops due to delays
would not be reported though. This commit adds counting of these stats
to the CPU factorization code. Counting is done just like in the GPU
code.