Avoid default-init overhead in assemble_pre map
With e3b35456 std::vector started to take care of the memory for the
lookup map in assemble_pre to avoid having to call allocate and
deallocate manually. Unfortunately, std::vector also default initializes
all values on construction, meaning that this change introduced an
overhead of always zeroing the map values initially.

Indeed it does not need to be zero'd, as all used values are set upon
building it. As it needs to be of size n to be able to map any row
index, zeroing it can actually be quite costly, especially since it is
repeated for each assemble_pre. Very noticeable for example on
DIMACS10/tx2010 from SuiteSparse matrix collection.

This replaces the std::vector with std::unique_ptr to avoid the repeated
unnecessary zeroing but still keep the implementation RAII.