Fix out of bounds std::vector subscript operator usage
In several places, the std::vector subscript operator[] is used
together with the addressof operator to obtain a pointer into the
vector. In some cases, this is intended to yield a pointer past the
end of the vector. In such cases, the pointer is not actually used
to read from or write to. Thus it causes no actual problems in
practice and even address sanitizer and valgrind are happy with it.
However, operator[] is still called with an out of bounds index,
which is undefined behaviour. This is correctly detected by
Microsofts standard library debug checks and also by GNU's library
once compiling with -D_GLIBCXX_DEBUG.

As a fix, we use .data() and pointer arithmetic to get to the
intended pointer without any out of bound access. Fixes #140