Use malloc for test matrices in SSMFE C tests
On Windows, the tests ran out of stack space with the many
variable-length arrays on the stack, so replacing with malloc. To still
be able to do all the matrix indexing, we use variably-modified types.
That requires dereferencing to be added in many places and makes the
syntax a bit unfamiliar, but it does avoid having to do the indexing
manually.