Avoid oversized arrays in scaling tests
Arrays for testing the scaling functions with random matrices were only
allocated once with the maximum dimensions of all possibly generated
matrices. That makes it simple, but unfortunately also has the potential
to hide possible bugs, as seen in #211 and #213. To improve the chance
to detect such bugs, this reallocates the arrays to the exact dimensions
for each current test case.