Fix buffer overflow in CPU block_ldlt in find_maxloc on NaNs
src/ssids/cpu/kernels/block_ldlt.hxx
(spral::ssids::cpu::block_ldlt_internal::find_maxloc): Add the same
sanity checks for (cloc < BLOCK_SIZE && rloc < BLOCK_SIZE) in the SIMD
case as in the non-SIMD case. This fixes a buffer overflow in Ipopt when
the input contains some NaNs. In that case, it can happen at this spot
that cbest and rbest are still at their initial
std::numeric_limits<int>::max() value, so cloc and rloc end up set to
that, and so a[cloc*lda+rloc] is out of bounds. The result will still be
garbage if we have NaNs, but at least SPRAL will not crash the entire
application anymore.
(spral::ssids::cpu::block_ldlt_internal::block_ldlt): Check whether the
found maxloc is valid (t and m both < BLOCK_SIZE), throw a SingularError
otherwise (which I think is a reasonable thing to claim when there were
NaNs in the input). Otherwise we eventually exit the entire application
with the pivsiz == 0 error case.